fix(test): Moved test functions into separate file from main.c.
Added TEST argument that runs regardless of hardware.
Added in CI with the build_tests.sh to automatically run all test.
