+++ Makefile
@@ -1,13 +1,22 @@
 #Check arguments
 ifeq ($(HW),LAUNCHPAD)
-TARGET_NAME=launchpad
+TARGET_HW=launchpad
 else ifeq ($(HW),NSUMO)
-TARGET_NAME=n_sumo
+TARGET_HW=n_sumo
 else ifeq ($(MAKECMDGOALS),clean)
 else ifeq ($(MAKECMDGOALS),cppcheck)
 else ifeq ($(MAKECMDGOALS),format)
 else
 $(error "Must pass HW=LAUNCHPAD or HW=NSUMO")
+endif
+TARGET_NAME=$(TARGET_HW)
+
+ifneq ($(TEST),) #TEST argument
+ifeq ($(findstring test_, $(TEST)),)
+$(error "TEST=$(TEST) is invalid (test function must start with test_)")
+else
+TARGET_NAME=$(TEST)
+endif
 endif
 
 # Directories
@@ -17,7 +26,7 @@ MSPGCC_BIN_DIR = $(MSPGCC_ROOT_DIR)/bin
 MSPGCC_INCLUDE_DIR = $(MSPGCC_ROOT_DIR)/include
 INCLUDE_DIRS = $(MSPGCC_INCLUDE_DIR)
 LIB_DIRS = $(MSPGCC_INCLUDE_DIR)
-BUILD_DIR = build/$(TARGET_NAME)
+BUILD_DIR = build
 OBJ_DIR = $(BUILD_DIR)/obj
 TI_CCS_DIR = $(TOOLS_DIR)/ccs2020/ccs
 DEBUG_BIN_DIR = $(TI_CCS_DIR)/ccs_base/DebugServer/bin
@@ -37,7 +46,7 @@ CPPCHECK = cppcheck
 FORMAT = clang-format
 
 # Files
-TARGET = $(BUILD_DIR)/$(TARGET_NAME)
+TARGET = $(BUILD_DIR)/bin/$(TARGET_HW)/$(TARGET_NAME)
 
 SOURCES_WITH_HEADERS = \
 	src/common/assert_handler.c \
@@ -48,22 +57,32 @@ SOURCES_WITH_HEADERS = \
 	src/drivers/io.c \
 	src/app/drive.c \
 	src/app/enemy.c \
-	
+
+ifndef TEST
 SOURCES = \
 	src/main.c \
 	$(SOURCES_WITH_HEADERS)
-
+else
+SOURCES = \
+	src/test/test.c \
+	$(SOURCES_WITH_HEADERS)
+#Delete object file to force rebuild when changing test
+$(shell rm -f $(BUILD_DIR)/obj/src/test/test.o)
+endif
 HEADERS = \
-	  $(SOURCES_WITH_HEADERS:.c=.h) \
-	  src/common/defines.h
+	$(SOURCES_WITH_HEADERS:.c=.h) \
+	src/common/defines.h
 
 OBJECT_NAMES = $(SOURCES:.c=.o)
 OBJECTS = $(patsubst %,$(OBJ_DIR)/%,$(OBJECT_NAMES))
 
 #Defines
-HW_DEFINES = $(addprefix -D, $(HW)) #e.g. -DSUMO or -DLAUNCHPAD
-DEFINES = $(HW_DEFINES)
-
+HW_DEFINE = $(addprefix -D, $(HW)) #e.g. -DSUMO or -DLAUNCHPAD
+TEST_DEFINE = $(addprefix -DTEST=, $(TEST)) 
+DEFINES = \
+	  $(HW_DEFINE) \
+	  $(TEST_DEFINE)
+	
 #Static Analysis - Skip over checking MSP430 helper headers due to checking every ifdefs...
 CPPCHECK_INCLUDES = ./src
 CPPCHECK_IGNORE = external/printf
